add sensor lib
