+++ lib/Sensor/Sensor.h
@@ -1,0 +1,6 @@
+#include <Adafruit_ADS1X15.h>
+#include <Arduino.h>
+
+int getSensorData();
+void switchSensor(bool on);
+void setupSensor();
