+++ src/main.cpp
@@ -1,18 +1,12 @@
 #include <Arduino.h>
 #include <WiFi.h>
 #include <WebSocketsServer.h>
-#include <Adafruit_ADS1X15.h>
 #include <ConfigManager.h>
+#include <Sensor.h>
 
-Adafruit_ADS1115 ads;  /* Use this for the 16-bit version */
-
-int getSensorData();
 void ledBlink();
-void switchSensor(bool on);
 
-const float slowResponseTime = 10.0;
 const int touchPin = 4; // GPIO4 as the touch-sensitive pin
-const int sensorControlPins[3] = {5, 18, 19};
 
 WebSocketsServer webSocket(81);
 
@@ -57,17 +51,8 @@ void setup() {
   Serial.begin(115200);
   checkWifiInfo();
 
-  Serial.println("ADC Range: +/- 6.144V (1 bit = 3mV/ADS1015, 0.1875mV/ADS1115)");
+  setupSensor();
 
-  if (!ads.begin()) {
-    Serial.println("Failed to initialize ADS.");
-    while (1);
-  }
-
-  for (int i = 0; i < 3; i++) {
-    pinMode(sensorControlPins[i], OUTPUT);
-  }
-  switchSensor(true);
   touchAttachInterrupt(touchPin, touchCallback, 40); // Attach touch interrupt
 
   //Configure Touchpad as wakeup source
@@ -98,35 +83,9 @@ void loop() {
   }
 }
 
-int getSensorData() {
-  int16_t adc0;
-  float volts0;
-
-  adc0 = ads.readADC_SingleEnded(0);
-
-  // volts0 = ads.computeVolts(adc0);
-
-  // Serial.println("-----------------------------------------------------------");
-  // Serial.print("AIN0: "); Serial.print(adc0); Serial.print("  "); Serial.print(volts0); Serial.println("V");
-
-  // response time
-  // fast: 1.5ms
-  // standard: 5ms
-  // slow: 10ms
-  delay(slowResponseTime);
-  return adc0;
-}
-
 // LED blink
 void ledBlink() {
   digitalWrite(LED_BUILTIN, HIGH);
   delay(500);
   digitalWrite(LED_BUILTIN, LOW);
-}
-
-// Switch sensor
-void switchSensor(bool on) {
-  for (int i = 0; i < 3; i++) {
-    digitalWrite(sensorControlPins[i], on ? HIGH : LOW);
-  }
 }
